Move some summer related code to a new files in the utils folder
